+++ modules/client-java/src/test/java/be/wegenenverkeer/atomium/japi/client/FailureTest.java
@@ -1,6 +1,6 @@
 package be.wegenenverkeer.atomium.japi.client;
 
-import be.wegenenverkeer.rxhttpclient.HttpClientError;
+import be.wegenenverkeer.atomium.japi.client.rxhttpclient.RxHttpAtomiumClient;
 import com.github.tomakehurst.wiremock.client.WireMock;
 import com.github.tomakehurst.wiremock.common.ClasspathFileSource;
 import com.github.tomakehurst.wiremock.common.Slf4jNotifier;
@@ -14,7 +14,6 @@ import org.junit.Test;
 import java.time.Duration;
 import java.util.concurrent.TimeUnit;
 
-import static be.wegenenverkeer.atomium.japi.client.FeedPositionStrategies.from;
 import static be.wegenenverkeer.atomium.japi.client.FeedPositionStrategies.fromNowOn;
 import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.wireMockConfig;
